requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
exceptiongroup==1.1.3
grpcio==1.46.3
grpcio-tools==1.46.3
idna==3.4
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
protobuf==3.20.3
PyAudio==0.2.13
pydub==0.25.1
pytest==7.4.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
tomli==2.0.1
urllib3==2.0.6
yandex-speechkit==1.3.3
